fix: resolve backend-size CI and remove accidental session artifact
Split oversized goal failure baseline and runtime_config files to satisfy the 1000-line backend-size gate, relocate blocking-load re-exports into config_validator_topology.go, and delete the durable-sessions file that was accidentally committed during the main merge.
